requirements.txt
appnope==0.1.3
asttokens==2.0.5
backcall==0.2.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.6
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
httplib2==0.20.4
idna==3.3
install==1.3.5
ipykernel==6.13.0
ipython==8.3.0
jedi==0.18.1
jupyter-client==7.3.1
jupyter-core==4.10.0
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
pywin32==306
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
stack-data==0.2.0
tornado==6.1
traitlets==5.2.0
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
